+++ apps/analog-app/src/app/app.config.ts
@@ -4,19 +4,20 @@ import {
   withInterceptors,
 } from '@angular/common/http';
 import { ApplicationConfig } from '@angular/core';
-import { provideClientHydration } from '@angular/platform-browser';
+import {
+  provideClientHydration,
+  withEventReplay,
+} from '@angular/platform-browser';
 import { provideFileRouter, requestContextInterceptor } from '@analogjs/router';
 import { withNavigationErrorHandler } from '@angular/router';
-
-import { cookieInterceptor } from './interceptors/cookies.interceptor';
 
 export const appConfig: ApplicationConfig = {
   providers: [
     provideFileRouter(withNavigationErrorHandler(console.error)),
     provideHttpClient(
       withFetch(),
-      withInterceptors([cookieInterceptor, requestContextInterceptor])
+      withInterceptors([requestContextInterceptor])
     ),
-    provideClientHydration(),
+    provideClientHydration(withEventReplay()),
   ],
 };